fix FloatSqlDescriptor so it can handle double-precision values
The SQL type FLOAT is *not* a single-precision type, at least not by
default. (It depends on the specified precision.)
