+++ hibernate-core/src/main/java/org/hibernate/type/descriptor/sql/spi/FloatSqlDescriptor.java
@@ -6,17 +6,42 @@
  */
 package org.hibernate.type.descriptor.sql.spi;
 
+import java.sql.CallableStatement;
+import java.sql.PreparedStatement;
+import java.sql.ResultSet;
+import java.sql.SQLException;
 import java.sql.Types;
 
+import org.hibernate.sql.AbstractJdbcValueBinder;
+import org.hibernate.sql.AbstractJdbcValueExtractor;
+import org.hibernate.sql.JdbcValueBinder;
+import org.hibernate.sql.JdbcValueExtractor;
+import org.hibernate.sql.exec.spi.ExecutionContext;
 import org.hibernate.type.descriptor.java.spi.BasicJavaDescriptor;
+import org.hibernate.type.descriptor.java.spi.JavaTypeDescriptor;
+import org.hibernate.type.descriptor.sql.internal.JdbcLiteralFormatterNumericData;
 import org.hibernate.type.spi.TypeConfiguration;
 
 /**
  * Descriptor for {@link Types#FLOAT FLOAT} handling.
  *
+ * The ANSI SQL {@link Types#FLOAT FLOAT} type is an approximate
+ * floating point type with configurable precision. Both
+ * {@link Types#DOUBLE DOUBLE} and {@link Types#REAL REAL} are
+ * flavors of {@link Types#FLOAT FLOAT} with hardcoded but
+ * implementation-specific precisions. The only constraint is
+ * that {@link Types#REAL REAL} is defined to be of lower
+ * precision than {@link Types#DOUBLE DOUBLE}. However, on some
+ * platforms these three types are all synonyms.
+ *
+ * By default we map this type to Java {@link Double}.
+ *
+ * By default we map both Java {@link Double} and {@link Float}
+ * to this type, but with distinct values for precision.
+ *
  * @author Steve Ebersole
  */
-public class FloatSqlDescriptor extends RealSqlDescriptor {
+public class FloatSqlDescriptor extends AbstractTemplateSqlTypeDescriptor {
 	public static final FloatSqlDescriptor INSTANCE = new FloatSqlDescriptor();
 
 	public FloatSqlDescriptor() {
@@ -28,7 +53,66 @@ public class FloatSqlDescriptor extends RealSqlDescriptor {
 	}
 
 	@Override
+	public boolean canBeRemapped() {
+		return true;
+	}
+
+	@Override
 	public <T> BasicJavaDescriptor<T> getJdbcRecommendedJavaTypeMapping(TypeConfiguration typeConfiguration) {
 		return (BasicJavaDescriptor<T>) typeConfiguration.getJavaTypeDescriptorRegistry().getDescriptor( Double.class );
+	}
+
+	@Override
+	@SuppressWarnings("unchecked")
+	public <T> JdbcLiteralFormatter<T> getJdbcLiteralFormatter(JavaTypeDescriptor<T> javaTypeDescriptor) {
+		return new JdbcLiteralFormatterNumericData( javaTypeDescriptor, Double.class );
+	}
+
+	@Override
+	protected <X> JdbcValueBinder<X> createBinder(BasicJavaDescriptor<X> javaTypeDescriptor, TypeConfiguration typeConfiguration) {
+		return new AbstractJdbcValueBinder<X>( javaTypeDescriptor, this ) {
+			@Override
+			protected void doBind(
+					PreparedStatement st,
+					int index, X value,
+					ExecutionContext executionContext) throws SQLException {
+				st.setDouble(
+						index,
+						javaTypeDescriptor.unwrap( value, Double.class, executionContext.getSession() )
+				);
+			}
+
+			@Override
+			protected void doBind(
+					CallableStatement st,
+					String name, X value,
+					ExecutionContext executionContext)
+					throws SQLException {
+				st.setDouble(
+						name,
+						javaTypeDescriptor.unwrap( value, Double.class, executionContext.getSession() )
+				);
+			}
+		};
+	}
+
+	@Override
+	protected <X> JdbcValueExtractor<X> createExtractor(BasicJavaDescriptor<X> javaTypeDescriptor, TypeConfiguration typeConfiguration) {
+		return new AbstractJdbcValueExtractor<X>( javaTypeDescriptor, this ) {
+			@Override
+			protected X doExtract(ResultSet rs, int position, ExecutionContext executionContext) throws SQLException {
+				return javaTypeDescriptor.wrap( rs.getDouble( position ), executionContext.getSession() );
+			}
+
+			@Override
+			protected X doExtract(CallableStatement statement, int position, ExecutionContext executionContext) throws SQLException {
+				return javaTypeDescriptor.wrap( statement.getDouble( position ), executionContext.getSession() );
+			}
+
+			@Override
+			protected X doExtract(CallableStatement statement, String name, ExecutionContext executionContext) throws SQLException {
+				return javaTypeDescriptor.wrap( statement.getDouble( name ), executionContext.getSession() );
+			}
+		};
 	}
 }
